refactor: rename Item file to QuestionOptionsItem and refactor references
